Added signin form component, fixed validation on submit, imroved error namings
