+++ src/modules/loginModule/loginModule.css
@@ -31,7 +31,7 @@
 }
 
 .login__form .button {
-    margin-top: 150px; 
+    margin-top: 120px; 
 }
 
 .login__form > :last-child {
