+++ src/modules/signinModule/signInModule.css
@@ -31,7 +31,7 @@
 }
 
 .signin__form .button {
-    margin-top: 30px; 
+    margin-top: 40px; 
 }
 
 .signin__form > :last-child {
