+++ src/utils/inputValidators/index.ts
@@ -1,0 +1,7 @@
+export {default as loginValidator} from './loginValidator.ts'
+export {default as passwordValidator} from './passwordValidator.ts'
+export {default as emailValidator} from './emailValidator.ts'
+export {default as firstNameValidator} from './firstNameValidator.ts'
+export {default as lastNameValidator} from './lastNameValidator.ts'
+export {default as phoneValidator} from './phoneValidator.ts'
+export {default as passwordCheckValidator} from './passwordCheckValidator.ts'
